+++ libs/components/src/lib/components/switcher/switcher.module.ts
@@ -1,14 +1,14 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { SwitcherComponent } from './switcher.component';
+import { PrizmSwitcherComponent } from './switcher.component';
 import { SwitcherItemComponent } from './components/switcher-item/switcher-item.component';
 import { PrizmIconModule } from '../icon';
 import { PrizmButtonModule } from '../button';
 import { PolymorphModule } from '../../directives';
 
 @NgModule({
-  declarations: [SwitcherComponent, SwitcherItemComponent],
+  declarations: [PrizmSwitcherComponent, SwitcherItemComponent],
   imports: [CommonModule, PrizmIconModule, PrizmButtonModule, PolymorphModule],
-  exports: [SwitcherComponent],
+  exports: [PrizmSwitcherComponent],
 })
 export class PrizmSwitcherModule {}